current API testing done

+++ routes/approveRegistration.js
@@ -10,10 +10,13 @@ const Ref = require("../models/user/ref");
 
 router.post("/", async (req,res) => {
       const approvalInfo = req.body;
+  console.log(approvalInfo);
       const final = {
-        user: approvalInfo.userId,
+        user: approvalInfo.user,
         status: approvalInfo.status,
       }
+
+
 
       try {
         const approval = new RegApproval(final);
@@ -23,10 +26,19 @@ router.post("/", async (req,res) => {
             throw new Error("approval info not added");
         }
 
+        const x = await User.find({_id: newApproval.user});
+        if(x.length === 0){
+          throw new Error({message: `user with id ${newApproval.user} not found in the database` });
+        }
+        console.log({message: `new approval: ${newApproval._id} created successfully`});
+       
 
         if(newApproval.status === "accept") {
-            const user = new RegisteredUser(newApproval.user);
-            const registeredUser = await user.save();
+          console.log("this is accepted status");
+            const x  = new RegisteredUser({
+              user: newApproval.user,
+            });
+            const registeredUser = await x.save();
 
 
             if(registeredUser === null) {
@@ -34,23 +46,31 @@ router.post("/", async (req,res) => {
             }
 
             // add registered User to the his reference
-           const ref = await Ref.update({_id: registeredUser.reference}, {$push: {refFrom: registeredUser._id}});
+          //  const ref = await Ref.update({_id: registeredUser.reference}, {$push: {refFrom: registeredUser._id}});
 
-           if(ref === null) {
-           throw new Error("ref not added");
-        }
+        //    if(ref === null) {
+        //    throw new Error("ref not added");
+        // }
         }
 
-        else {
-            const user = new RejectedUser(newApproval.user);
+        else if(newApproval.status === "reject") {
+          console.log('this is rejected status');
+            const user = new RejectedUser({
+              user: newApproval.user
+            });
             const rejectedUser = await user.save();
+            if(rejectedUser === null) {
+              throw new Error("rejected user could not be added");
+            }
 
              const deleted_user = await User.deleteOne({_id: newApproval.user});
              if(deleted_user!== null) {
-                res.status(200).json({message: "User deleted successfully"});
+              console.log("User deleted successfully");
+                // res.status(200).json({message: "User deleted successfully"});
              }
              else {
-              res.json({message: "user could not be deleted.."});
+              console.log("user could not be deleted");
+              // res.json({message: "user could not be deleted.."});
              }
         }
 
@@ -58,7 +78,7 @@ router.post("/", async (req,res) => {
       }
 
       catch(err) {
-console.log({reg_approval_route: err.message})
+console.log({message: err.message});
       }
